Time to stop playing around with this one and ship it home.
Added a decorator to check on coordinate values while I was messing with incoming data.  That slows down the execution so it's disabled from use

Separated functions out of the main file into support files to clean up the workflow.

Removed a bunch of asserts in favor of the decorator approach ... however, that's gone now.

Updated readme with some final notes and instructions for execution (matches requirements)